+++ src/constants/queryKeys.js
@@ -1,2 +1,1 @@
 export const QUERY_KEY_AUTHORITIES = 'authorities';
-export const QUERY_KEY_AUTHORITY_SOURCE = 'authority-source';